feat: building, unit 서비스 클래스에 validation 추가, ID 파라미터 null 체크 추가

+++ src/main/java/com/dobot/imjang/domain/unit/UnitService.java
@@ -1,6 +1,7 @@
 package com.dobot.imjang.domain.unit;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
 import org.springframework.stereotype.Service;
@@ -15,6 +16,8 @@ import jakarta.transaction.Transactional;
 
 @Service
 public class UnitService {
+    private static final String UNIT_ID_SHOULD_NOT_BE_NULL = "Unit id should not be null.";
+    private static final String BUILDING_ID_SHOULD_NOT_BE_NULL = "Building id should not be null.";
     private final UnitRepository unitRepository;
     private final BuildingRepository buildingRepository;
     private final UploadResultRepository uploadResultRepository;
@@ -31,6 +34,9 @@ public class UnitService {
     }
 
     public Unit createUnit(UnitCreateOrUpdateDto dto, UUID buildingId) {
+        if (Objects.isNull(buildingId)) {
+            throw new CustomException(ErrorCode.INVALID_INPUT, BUILDING_ID_SHOULD_NOT_BE_NULL);
+        }
         Building building = this.buildingRepository.findById(buildingId)
                 .orElseThrow(() -> new CustomException(ErrorCode.BUILDING_NOT_FOUND));
 
@@ -44,6 +50,9 @@ public class UnitService {
 
     @Transactional
     public Unit updateUnit(UUID id, UnitCreateOrUpdateDto dto) {
+        if (Objects.isNull(id)) {
+            throw new CustomException(ErrorCode.INVALID_INPUT, UNIT_ID_SHOULD_NOT_BE_NULL);
+        }
         Unit unit = this.unitRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.UNIT_NOT_FOUND));
         Unit updatingUnit = this.addUnitProperties(dto, unit);
         return this.unitRepository.save(updatingUnit);
@@ -51,6 +60,9 @@ public class UnitService {
 
     @Transactional
     public void deleteUnitById(UUID id) {
+        if (Objects.isNull(id)) {
+            throw new CustomException(ErrorCode.INVALID_INPUT, UNIT_ID_SHOULD_NOT_BE_NULL);
+        }
         this.unitRepository.deleteById(id);
     }
